[CIR-131]: add models and tests for full return

+++ app/models/fullReturn/AllocatedReactivationsModel.scala
@@ -18,10 +18,10 @@ package models.fullReturn
 
 import play.api.libs.json.Json
 
-case class AllocatedReactivationsModel(ap1NetDisallowances: Int,
-                                       currentPeriodReactivation: Int,
-                                       totalReactivation: Int,
-                                       reactivationCap: Int)
+case class AllocatedReactivationsModel(ap1NetDisallowances: BigDecimal,
+                                       currentPeriodReactivation: BigDecimal,
+                                       totalReactivation: BigDecimal,
+                                       reactivationCap: BigDecimal)
 
 object AllocatedReactivationsModel {
   implicit val format = Json.format[AllocatedReactivationsModel]
